fix to set on the primary monitor, not just laptop display
This update provides the ability for PingPlace to move the notification
pop up on the primary display. For example, if you have a laptop
connected to external monitors with the laptop display closed, it
defaults to the top-right corner. This will check for which display is
acting as the primary monitor (ie where the dock is). In my testing, my
calendar widget was also causing issue so there's a check for that.